requirements.txt
streamlit==1.42.0
langgraph==0.2.6
langchain-core==0.3.2
langchain-openai==0.2.1
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
pypdf==5.4.0
python-dotenv==1.1.0
openai==1.75.0
ollama==0.4.8
groq==0.11.0
